APL-1598 Move out inner runable classes GetMoreBlocksThread and RestorePrunableDataTask from the BlockchainProcessorImpl.
